requirements.txt
openai>=1.0.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
streamlit>=1.28.0
python-dotenv>=1.0.0
httpx>=0.24.0